Move SortMode and SortOrder into their own files
as they are not specific to FieldSort
